requirements.txt
asgiref==3.7.2
Django==5.0.3
django-filter==24.2
djangorestframework==3.15.1
Markdown==3.6
mysqlclient==2.2.4
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
